saluki_common::buf

Struct ChunkedBytesBuffer

pub struct ChunkedBytesBuffer { /* private fields */ }

A bytes buffer that write dynamically-sized payloads across multiple fixed-size chunks.

As callers write data to ChunkedBytesBuffer, it will allocate additional chunks as needed, and write the data across these chunks. This allows for predictable memory usage by avoiding reallocations that overestimate the necessary additional capacity, and provides mechnical sympathy to the allocator by using consistently-sized chunks.

ChunkedBytesBuffer implements AsyncWrite and Body, allowing it to be asynchronously written to and used as the body of an HTTP request without any additional allocations and copying/merging of data into a single buffer.

Implementations

impl ChunkedBytesBuffer

pub fn new(chunk_size: usize) -> Self

Creates a new ChunkedBytesBuffer, configured to use chunks of the given size.

pub fn is_empty(&self) -> bool

Returns true if the buffer has no data.

pub fn len(&self) -> usize

Returns the number of bytes written to the buffer.

pub fn freeze(self) -> FrozenChunkedBytesBuffer

Consumes this buffer and returns a read-only version of it.

All existing chunks at the time of calling this method will be present in the read-only buffer.