saluki_common::buf

Struct FrozenChunkedBytesBuffer

pub struct FrozenChunkedBytesBuffer { /* private fields */ }

A frozen, read-only version of ChunkedBytesBuffer.

FrozenChunkedBytesBuffer can be cheaply cloned, and allows for sharing the underlying chunks among multiple tasks.

Implementations

impl FrozenChunkedBytesBuffer

pub fn is_empty(&self) -> bool

Returns true if the buffer has no data.

pub fn len(&self) -> usize

Returns the number of bytes written to the buffer.