saluki_common::hash

Function get_fast_build_hasher

pub fn get_fast_build_hasher() -> FastBuildHasher

Returns a fresh FastBuildHasher instance.

This instance should not be used to generate hashes that will be compared against hashes generated either with a hasher acquired from get_fast_hasher or hash_single_fast, and those methods use a shared global state to both speed up hashing and ensure that the hashes are consistent across runs of those functions in particular.