saluki_common::scrubber

Struct Scrubber

pub struct Scrubber { /* private fields */ }

A YAML scrubber that can be configured with different replacers.

Implementations

impl Scrubber

pub fn new() -> Self

Creates a new Scrubber with no replacers.

pub fn add_replacer(&mut self, replacer: Replacer)

Adds a replacer to the scrubber.

pub fn scrub_bytes(&self, data: &[u8]) -> Vec<u8>

Scrubs sensitive data from a byte slice.

This method will scrub the data, returning a new byte vector.