saluki_common::strings

Function lower_alphanumeric

pub fn lower_alphanumeric(s: &str) -> String

Sanitizes the input string by ensuring all characters are lowercase ASCII alphanumeric or underscores.

All characters that are not ASCII alphanumeric or underscores are replaced with underscores, and alphanumerics will be lowercased.