saluki_common::strings

Struct StringBuilder

pub struct StringBuilder<I = ()> { /* private fields */ }

A string builder.

This builder is designed to allow building strings incrementally. This can simplify certain patterns of string construction by removing the need to manually manage a temporary string buffer, clearing it after building the resulting string, and so on.

Limits

The builder can be configured to limit the overall length of the strings it builds.

Interning

The builder supports providing an interner that is used to intern the finalized string. This allows for efficiently building strings, reusing the intermediate buffer in between before eventually interning the string.

Implementations

impl StringBuilder<()>

pub fn new() -> Self

Creates a new StringBuilder.

No limit is set for the strings built by this builder.

pub fn with_limit(limit: usize) -> Self

Creates a new StringBuilder with the given limit.

Strings that exceed the limit will be discarded.

impl<I> StringBuilder<I>

pub fn with_interner<I2>(self, interner: I2) -> StringBuilder<I2>
where I2: Interner,

Configures this builder with the given interner.

pub fn is_empty(&self) -> bool

Returns true if the buffer of the builder is empty.

pub fn len(&self) -> usize

Returns the length of the buffer of the builder.

pub fn available(&self) -> usize

Returns the available space in the buffer of the builder.

pub fn clear(&mut self)

Clears the buffer of the builder.

pub fn push(&mut self, c: char) -> Option<()>

Pushes a character into the builder.

Returns None if the buffer limit would be exceeded by writing the character.

pub fn push_str(&mut self, s: &str) -> Option<()>

Pushes a string fragment into the builder.

Returns None if the buffer limit would be exceeded by writing the string.

pub fn as_str(&self) -> &str

Returns a references to the current string.

impl<I> StringBuilder<I>
where I: Interner,

pub fn try_intern(&mut self) -> Option<MetaString>

Attempts to build and intern the string.

Returns None if the string exceeds the configured limit or if it cannot be interned.